LayoutTests/fast/parser/foster-parent-expected.txt

PASS div.children.length is 0
PASS docFragment.firstElementChild.tagName is "H1"
PASS successfullyParsed is true

TEST COMPLETE
